requirements.txt
pandas
fastapi
uvicorn[standard]
pymongo
certifi
python-dotenv